requirement.txt
Django==5.2.6
mysqlclient==2.2.7
pillow==11.3.0
pyfiglet==0.8.post1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1